requirements.txt
fastapi
python-multipart
rembg
Pillow
uvicorn